Fixes compilation issue and updates c interface
The compilation issues was resolved by switching to the clang compiler
and compiling for the x86_64 architecture. This makes no sense, but it
does work, so moving on ...

I also modified the C interface to match the most recent Python
interface.

WARNING: This code appears to have a few bugs, and produces incorrect
results.